refactor operations to a separate file from transactions
