+++ quadratic-core/src/controller.rs
@@ -7,8 +7,10 @@ use self::transactions::Transaction;
 
 pub mod cells;
 pub mod clipboard;
+pub mod compute;
 pub mod dependencies;
 pub mod formatting;
+pub mod operations;
 pub mod sheets;
 pub mod transactions;
